Refactor of serializers to multiple files.
